+++ CHANGELOG.md
@@ -6,6 +6,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.2.0] - 2022-12-20
+### Added
+- support for `async` callbacks
+- `off()` methods to remove listeners of an event
+
+### Changed
+- Not using node `events` module anymore
+
 ## [0.1.0] - 2021-10-06
 ### Added
 - Events class